refactor: Make it can be used as a library
